requirements/requirements-dev.txt
# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
behave==1.2.6
allure-behave==2.13.2
allure-python-commons==2.13.2

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Development Tools
pre-commit==3.6.0
tox==4.13.0
twine==4.0.2

# IDE Support
python-language-server==0.36.2
pylint==3.0.3

# Data Quality
great-expectations==0.18.11
pandas==2.1.4
numpy==1.26.3
sqlalchemy==2.0.25 